feat: mode AUTH methods out from USERS routes
* create AUTH own module
